# backend_backup_20250927_184832/HR_HELPER_COMPLETE_API_SPECIFICATION.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HR Helper Complete API",
    "description": "Полная спецификация API системы HR Helper - комплексной платформы для управления HR процессами, включающей управление пользователями, финансами, вакансиями, интервьюерами, AI сервисы и интеграции с внешними системами.",
    "version": "2.0.0",
    "contact": {
      "name": "HR Helper Development Team",
      "email": "support@hrhelper.com",
      "url": "https://hrhelper.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.hrhelper.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "SessionAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/vacancies/{id}/toggle-active/": {
      "post": {
        "summary": "Переключение активности вакансии",
        "description": "Активация или деактивация вакансии",
        "tags": ["Vacancies"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "Статус вакансии изменен",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Vacancy"}
              }
            }
          }
        }
      }
    },
    "/api/vacancies/{id}/assign-grades/": {
      "post": {
        "summary": "Назначение грейдов вакансии",
        "description": "Назначение грейдов для вакансии",
        "tags": ["Vacancies"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "grade_ids": {
                    "type": "array",
                    "items": {"type": "integer"}
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Грейды назначены",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Vacancy"}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Vacancy": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string"},
          "external_id": {"type": "string"},
          "recruiter": {"type": "integer"},
          "recruiter_name": {"type": "string"},
          "invite_title": {"type": "string"},
          "invite_text": {"type": "string"},
          "scorecard_title": {"type": "string"},
          "scorecard_text": {"type": "string"},
          "available_grades": {
            "type": "array",
            "items": {"type": "integer"}
          },
          "available_grades_names": {
            "type": "array",
            "items": {"type": "string"}
          },
          "is_active": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
...
    },
    "securitySchemes": {
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}